+++ RELEASE-NOTES.md
@@ -9,6 +9,10 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 
 ## [Unreleased] - Release date yyyy-mm-dd
 
+### Changed
+- Modified `blt_add_sphinx_target` to allow setting sphinx source and configure directories.
+- Modified `blt_add_sphinx_target` and `blt_add_doxygen_target` requiring the first argument preceed with TARGET.
+
 ## [Version 0.7.2] - Release date 2026-05-01
 
 ### Changed